requirements.txt
blinker==1.6.2
certifi==2022.12.7
click==8.1.3
colorama==0.4.6
Flask==2.3.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0
sib-api-v3-sdk==7.6.0
six==1.16.0
SQLAlchemy==2.0.11
SQLAlchemy-serializer==1.4.1
typing_extensions==4.5.0
urllib3==2.0.1
waitress==2.1.2
Werkzeug==2.3.2
WTForms==3.0.1
zipp==3.15.0
python==3.8.16